test: 재사용 가능한 test fixture를 별도 클래스로 분리
