+++ src/test/java/nextstep/subway/line/acceptance/testfactory/LineAcceptanceTestFactory.java
@@ -1,0 +1,125 @@
+package nextstep.subway.line.acceptance.testfactory;
+
+import static org.assertj.core.api.Assertions.*;
+
+import java.util.List;
+import java.util.stream.Collectors;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+
+import io.restassured.RestAssured;
+import io.restassured.response.ExtractableResponse;
+import io.restassured.response.Response;
+import nextstep.subway.line.dto.LineRequest;
+import nextstep.subway.line.dto.LineResponse;
+import nextstep.subway.station.dto.StationResponse;
+
+public class LineAcceptanceTestFactory {
+
+	public static ExtractableResponse<Response> 지하철_노선_등록되어_있음(LineRequest params) {
+		return 지하철_노선_생성_요청(params);
+	}
+
+	public static ExtractableResponse<Response> 지하철_노선_등록되어_있음(String name, String color, StationResponse upStation, StationResponse downStation, int distance) {
+		return 지하철_노선_생성_요청(new LineRequest(name,color,upStation.getId(),downStation.getId(),distance));
+	}
+
+	public static ExtractableResponse<Response> 지하철_노선_생성_요청(LineRequest params) {
+		return RestAssured
+			.given().log().all()
+			.contentType(MediaType.APPLICATION_JSON_VALUE)
+			.body(params)
+			.when().post("/lines")
+			.then().log().all().
+				extract();
+	}
+
+	public static ExtractableResponse<Response> 지하철_노선_목록_조회_요청() {
+		return 지하철_노선_목록_조회_요청("/lines");
+	}
+
+	public static ExtractableResponse<Response> 지하철_노선_목록_조회_요청(ExtractableResponse<Response> response) {
+		String uri = response.header("Location");
+
+		return 지하철_노선_목록_조회_요청(uri);
+	}
+
+	private static ExtractableResponse<Response> 지하철_노선_목록_조회_요청(String uri) {
+		return RestAssured
+			.given().log().all()
+			.accept(MediaType.APPLICATION_JSON_VALUE)
+			.when().get(uri)
+			.then().log().all()
+			.extract();
+	}
+
+	public static ExtractableResponse<Response> 지하철_노선_조회_요청(LineResponse response) {
+		return RestAssured
+			.given().log().all()
+			.accept(MediaType.APPLICATION_JSON_VALUE)
+			.when().get("/lines/{lineId}", response.getId())
+			.then().log().all()
+			.extract();
+	}
+
+	public static ExtractableResponse<Response> 지하철_노선_수정_요청(ExtractableResponse<Response> response, LineRequest params) {
+		String uri = response.header("Location");
+
+		return RestAssured
+			.given().log().all()
+			.contentType(MediaType.APPLICATION_JSON_VALUE)
+			.body(params)
+			.when().put(uri)
+			.then().log().all()
+			.extract();
+	}
+
+	public static ExtractableResponse<Response> 지하철_노선_제거_요청(ExtractableResponse<Response> response) {
+		String uri = response.header("Location");
+
+		return RestAssured
+			.given().log().all()
+			.when().delete(uri)
+			.then().log().all()
+			.extract();
+	}
+
+	public static void 지하철_노선_생성됨(ExtractableResponse response) {
+		assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
+		assertThat(response.header("Location")).isNotBlank();
+	}
+
+	public static void 지하철_노선_생성_실패됨(ExtractableResponse<Response> response) {
+		assertThat(response.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
+	}
+
+	public static void 지하철_노선_목록_응답됨(ExtractableResponse<Response> response) {
+		assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
+	}
+
+	public static void 지하철_노선_응답됨(ExtractableResponse<Response> response, ExtractableResponse<Response> createdResponse) {
+		assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
+		assertThat(response.as(LineResponse.class)).isNotNull();
+	}
+
+	public static void 지하철_노선_목록_포함됨(ExtractableResponse<Response> response, List<ExtractableResponse<Response>> createdResponses) {
+		List<Long> expectedLineIds = createdResponses.stream()
+			.map(it -> Long.parseLong(it.header("Location").split("/")[2]))
+			.collect(Collectors.toList());
+
+		List<Long> resultLineIds = response.jsonPath().getList(".", LineResponse.class).stream()
+			.map(LineResponse::getId)
+			.collect(Collectors.toList());
+
+		assertThat(resultLineIds).containsAll(expectedLineIds);
+	}
+
+	public static void 지하철_노선_수정됨(ExtractableResponse<Response> response) {
+		assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
+	}
+
+	public static void 지하철_노선_삭제됨(ExtractableResponse<Response> response) {
+		assertThat(response.statusCode()).isEqualTo(HttpStatus.NO_CONTENT.value());
+	}
+}
